setuptools: Fix description, license tag, and install package entries
The description is a bit weird and does not exactly describe what
this tool does. This is an attempt to better describe what it does.

Also, adjust the license name to match what the code actually says it
is: GPLv3+.

Additionally, this script installs stuff that should not be installed
because of its usage of 'find_packages()'. Since there's only one
actual package that needs to be installed, we just list it now.