Do not rename .repo files
Currently, when using RHSM, convert2rhel renames all non-RHSM repofiles to
make sure only the RHSM ones are used. That causes a problem when somebody
passes --enablerepo together with using RHSM, and this repo is in some
other repofile than /etc/yum.repos.d/redhat.repo. Because of the repofile
renaming, this repository is then unexpectedly not used during the
conversion.

Co-authored-by: Eric Gustavsson <SpyTec@users.noreply.github.com>