Do not expect offline snapshot of RHEL repos
The feature of consuming the offline snapshot of RHEL repos hasn't ever been
used in upstream and it's not planned to be ever used. It used to provide
information to the user prior the conversion about what packages *might*
not be available in RHEL. Considering that we have no mechanism to get such
snapshot publicly, the feature is being removed to make the code simpler.