Do not require python-syspurpose in spec (#117)
The python-syspurpose and python3-syspurpose packages are the required
dependencies of the RHEL 7 and 8 subscription-mananger package,
respectively. However this syspurpose package is not available on Oracle
Linux, so on OL 7 and 8, the installation of convert2rhel is not
possible, unless one installs the syspurpose package from a non-standard
repository manually beforehand.
Since the main and documented use of convert2rhel is currently with
using custom repositories, and not RHSM, we are removing this
dependency.
If anyone would like to use subscription-manager, they need to download
the syspurpose package to the /usr/share/convert2rhel/subscription-manager/
upon installing convert2rhel, alongside the subscription-manager pkgs.