Do not rename repofiles
We currently rename all repofiles not coming from Red Hat during the
conversion so that they are not used for the conversion in any way.
That is not needed if we call yum with the right parameters
(--enablerepo and --disablerepo). Using these parameters has been done
under https://github.com/oamg/convert2rhel/pull/84.