Remove CentOS/EPEL 6 from CI/CD
EPEL 6 is not available anymore. Not available in Fedora Copr as chroot,
not available as a repository.
As well, CentOS 6 repos are not available anymore, except Vault which is
an archive of old and not updated versions of CentOS.

That impacts our pylint tests, package building, unit testing. We'll be
looking for another ways to have the packages for CentOS/OL 6 built and
the code tested for py2.6 compatibility.