Remove local centos{6,7} requirements
In order to install a local dev env for centos 6 and 7, it is necessary still
run it inside centos6,7 systems (or similar). Otherwise the local environment
is incompatible and requires some intervention to make it work.

Considering that the local dev env is a compliment for the developer and
tehre is still a robust container way to run tests. We decided to leave only py3 env
as a local one.