Remove repofile pkgs after submgr installation
-> We need the original system repositories available in order to have
access to the dependencies of subscription-manager when we get to
installing it with yum/dnf. For that reason we need to make sure the
packages containg the system yum/dnf repofiles are removed only after
installing subscription-manager.

Remove release pkgs only after installing submgr and subscribing the
system -> we were removing the release pkgs before installing the
subscription-manager pkgs and we were passing the --releasever option to
yum with a value for RHEL repos, not the original system repos. That
caused the submgr installation to fail. And on CentOS 8, subscription
of the system fails if there's no release package installed
(centos-linux-release, redhat-release).