Use RHSM repos for downloading same version kernel
For the case when the all installed kernels have the same version as
those available in RHEL repos, we are forced to download one of the
RHEL kernels and force replace the already installed one. However for
the download we were using only the RHEL repos passed through the
--enablerepo option. That would cause the download to fail if the tool
uses RHSM and not custom repos (--enablerepo).