Allow caplog see logger entries
Caplog adds own handler to the root logger to track the records, however
our logger setup blocks the handler from being propogated to all child loggers.
This makes the caplog blind to the emitted entries.

Also tests are rewritten in pytest, because the old way was tested the logger
indirectly and switch to propagate=True leaded to inconsistent results across
different python versions.