Do not check deps when force replacing kernel
Under typical circumstances, kernel is installed using yum and with that the
system-release/redhat-release pkg is installed as a dependency. However when
the original OR kernel and RHEL kernel have the same version, we install it
through rpm which causes the "system-release is needed by kernel" error
because system-release/redhat-release is not installed at that point.