Inhibit the system conversion on UEFI
In case the system is booted using the UEFI firmware, the bootloader
is broken (user ends with the grub prompt). Inhibit the conversion
in case the UEFI is detected to prevent users from the broken
bootloader state.

Note: The current UEFI detection solution is not handling hybrid boot
(e.g. used on AWS, Azure, ...). It's possible that in case of hybrid
boot an additional checks will be needed. But such systems are not
supported for conversions right now so using just the simple detection.

Relates:  https://bugzilla.redhat.com/show_bug.cgi?id=1898314

Co-author: Fellipe Henrique <fellipeh@gmail.com>

v1: applied suggested changes in logged messages and tests