Unregister the system before removing sub-man
Before installing the subscription-manager packages known to work
fine with Red Hat CDN/Satellite, the convert2rhel needs to remove the
previously installed subscription-manager packages. That can lead to problems
in case the system was already registered. We should be unregistering the
system first before removing the sub-man pkgs.

The katello consumer rpm is very much RHSM-related. We need to remove it
alongside the subscription-manager. Without doing so, any previously
installed katello package would block the installation of the katello
package wanted for the conversion. And even if the wanted one is the
same, it needs to be installed together with (or after) the installation
of the subscription-manager, otherwise the RHSM config would not be
updated as necessary.

This action is not possible to revert during the rollback though.