Remove anaconda-dracut (in the optional repo)
The anaconda-dracut and anaconda-core rpms share by mistake 2 libexec
libraries on RHEL 7. As the anaconda-dracut rpm is present just inside
the optional repository on RHEL 7 which is by default disabled for
the conversion, the rpm transaction ends with conflicts between
anaconda-dracut and anaconda-core rpms. It do not even seems expected
that users could want to use the rpm at all. So just remove the rpm
during the conversion if installed.