Add check function to validate custom repos files and invalid custom repo param
When using custom repos but the repoid passed through --enablerepo is not valid (typo in the repoid, the baseurl not accessible, etc.),
convert2rhel continues without a problem beyond the point of no return (PONR) and then it fails, leaving the user with a message that
they need to fix their system manually.

This PR check the validity os repos before anything.

Apply suggestions from code review

Co-authored-by: Michal Bocek <mbocek@redhat.com>