Resolve broken rollback on OL 8 with RHSM
During the rollback on Oracle Linux 8 when RHSM was used for the
conversion, we are removing the subscription-manager packages installed
at the beginning of the conversion. For that we use `yum remove` and
that by default removes also all unused dependencies.

In this particular case it was removing also the usermode package, but
that is needed by the rhn-setup package that we need to install as part
of the rollback one step later using `rpm -i`. That installation fails
on the missing usermode dependency.

The solution is to use the `--noautoremove` yum option to not
automatically remove the no longer used deps. This solution has the
downside that the system is left with some installed packages that were
not installed before running convert2rhel (the dependencies of
subscription-manager).