Remove non-interactive command suggestion
Currently the feature of suggesting a convert2rhel command to run on other
systems to convert in a non-interactive fashion is not working properly.
For example, it is missing couple of switches (e.g. --no-rpm-va and --debug)
and does not handle well the situation with enabled repoids when RHSM is used.

To reduce the codebase to maintain and for the lack of feedback on this
feature, we decided to remove it.