Improve messaging around the unsupported kmods error
The message around unsupported kernel modules was not asking to make
sure the system runs the latest kernel which is important prerequisite
of the kernel module check.

Also, it was stating that disabling the modules will help. But it does
not. The module needs to be removed in order to not be picked up by the
check.

And printing plainly the `uname -r` call result was confusing.