chore: allow disabling of captured logs
Whenever we get an error with pytest it displays the captured logs for
stdout and stderr. This may flood the terminal with a lot of logs which
is difficult to parse.

To fix this we disable the capturing of logs by default and add an option
for adding logs using `SHOW_CAPTURE=yes make tests` instead.