Consolidate handling copr config
For `make copr-build` we expected .copr.conf file in the repo root but the
scripts/tft expected the copr config to be in ~/.config/copr (the
copr-cli default) when running `copr-cli status`.

Also the scripts/tft wasn't expecting any STDERR output of `make
copr-build` but it gives this output everytime so now we check the exit
code of `make copr-build` instead.