Avoid mutable types as default args.
This change removes two places where we use a set as a default value.
Mutable containers should not be used as default arguments as they are
created when the function is defined (basically, the first time the
module is imported) and then that one instance is used anytime the
method is invoked without that argument specified.  This means that
changes to the container will persist to the next time the function is
invoked which could lead to bugs down the road.

I do not believe we're currently experiencing any bugs due to this as we
never seem to mutate the containers.  Instead, we use the data to make
new sets inside of the functions.