Port test_download_rhsm_pkgs from unittest to pytest.
This way we can use parametrize to separate each call of the function
we're testing.  Previously, when it was all done in a single function,
the test could have side-effects carried over from one call to the
next.  For instance, in the test, the count of how many times
_download_rhsm_pkgs() was called incremented on every call to the public
function.  Using parametrize resets that counter to zero between every
run which allows the test's code to express the logical constraints
(_download_rhsm_pkgs() is called once) rather than an implementation
(_download_rhsm_pkgs() was called once more than the number of times it
was called previously).