Port run_subprocess and run_cmd_in_pty to take the command as a list instead of a string.
Sending a subprocess command as a string can be prone to bugs because
the user has to remember to quote any value that could have a space in
it.  When the values are stored in variables, it can be hard to remember
to do that.  If you don't quote, then arguments might be inappropriately
split when a space occurs in the user supplied data.

Passing the command as a list instead avoids this because the programmer
has to consider what are separate arguments when they place each element
in the list.  Placing a variable directly into the list will alway be
interpreted as a single argument even if there are embedded spaces.

For now, also check that we aren't misusing the cmd argument by
passing in a list anywhere in the code.

Fixes: https://issues.redhat.com/browse/OAMG-6183