Call `check_custom_repos_are_valid` before any other checks
This change intends to call the `check_custom_repos_are_valid` function
as the first thing before any other checks. The reason for that is we
can guarantee that we will work with the most recent cache for
`yum`/`dnf`.

Jira reference (If any): https://issues.redhat.com/browse/OAMG-6294
Bugzilla reference (If any):