Use ensurepip to install pip for testing.
Since we're only installing pip for python3, go ahead and use ensurepip
to install it.  That way we don't have to rely on a separate script from
the internet.

Also switch from the shell to command ansible module.  Shell passes the
comand to the shell so it can handle env vars, pipelines, redirection,
and other shell features.  However, that also means that shell special
characters have to be escaped.  So it's a good idea to use command
unless shell features are specifically needed.