Add condition to not backup specific package
Right now we have a package that exists in CentOS Linux 8.4, but was not
ported to CentOS Linux 8.5 causing a conversion error when that package
is present on any system.

The simplest fix to this is not backing up the package when we remove it
as the backup process uses `yumdownloader` and therefore,
`yumdownloader` uses any enabled repos that are present on the system.

Jira reference (If any): https://issues.redhat.com/browse/OAMG-6422
Bugzilla reference (If any): https://bugzilla.redhat.com/show_bug.cgi?id=2046292

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>