Fix lazy evaluation of string logs for loglevel.FILE
For log messages, you can do lazy evaluation of the format string to
occassionally boost performance when a log message is not used.  For
instance:

  class Hypothetical(object):
      def __str__(self):
          return self.calculate_fibonacci_10000()
  log.debug("10,000th fibonacci number is: %s", Hypothetical())

The format string will not be interpolated (and thus the 10,000th
fibonacci number will not be calculated) unless we are actually going to
log the debug() output.

The code in commit b870460 broke that because it calculated its own
format string and args for the message, interfering with whatever was
passed in via function parameter.  This change fixes that by moving the
formatting pieces of b870460 into the CustomFormatter class.  By that
time, the message string has already been interpolated so any changes to
the message string will not interfere.  Additionally, in the formatter
class, those pieces can be performed by standard logging.Formatter
features so we don't need the custom code to do that anymore.

OAMG-6571 #in-progress