Fix arguments to run_subprocess to be a list of strings.
One of the arguments to the command was an int instead of a string.
This didn't cause a problem for run_subprocess() itself but caused
loggerinst.debug() to error when it tried to format the log message
using the argument list.  This change will convert the int to a string
before calling run_subprocess()