Refactor the restorable/backup classes
Refactoring the restorable/backup classes and moving them from `utils.py` to
`backup.py` for improved readability.

The file `utils.py` had a circular dependency import with `systeminfo.py`, and
since we needed to recreate a path of the hadcoded repofiles, it was easier to
refactor and move it to a separate file instead of importing the
`systeminfo.py` inside a function that can be called multiple times and
re-import the same thing over and over.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>