Set terminal size before pexpect spawns the process.
Certain programs like yumdownloader need to have the terminal size set
before they start.  The version of pexpect in RHEL8 has a way,
pexpect.spawn()'s dimensions parameter, to set the terminal size at that
time.  We should make use of it there.

On RHEL7, pexpect doesn't have an official way to set the window size
but we have been using a workaround which involves overriding
setwinsize() to use different values since the RHEL7 version of pexpect
calls setwinsize() to set its own hardcoded values.  This patch cleans
that usage up slightly so that setwinsize is overridden when __init__()
is run but reverted back to its normal value afterwards.

RHELC-570 #in-progress