Auto-attach subscription when there is only one (#385)
Also, previous to this change, if a pool had been specified on the
convert2rhel command line and attach_subscription() was called more than
once, the subscription-manager command line would be constructed
incorrectly.  It would have "--pool" added to the front of the pool id
each time attach-subscription was called leading to something like this:
"subscription-manager attach --pool --pool POOLID".

To fix this, we now treat tool_opts.pool as read-only and do not modify
it on each call to attach_subscription().  This goes along with our
overall move to treat global singletons as read-only whenever possible.