Fix old logs archival
This PR introduces a fix for the log archival introduced in the
[PR#359](https://github.com/oamg/convert2rhel/pull/359).

The feature introduced there did a `copyfile` instead of a `move` to the
`archive` directory, wich was not the expected way of archiving a log.

Using a `copyfile` instead of a `move` causes the log file to be copied to the
`archive` directory as expected, but we still keep appending the log to the
`convert2rhel.log` file.

Jira reference (If any): https://issues.redhat.com/browse/RHELC-566
Bugzilla reference (If any):

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>