Map multiple TF distros to one copr build
The Fedora Copr is going to remove Oracle Linux chroots and we'll be
left with EPEL chroots only. However we can use the an EPEL copr build
for tests on both CentOS Linux and Oracle Linux.
This change removes our dependence on the soon-to-be-removed oraclelinux
chroots.

Related: https://github.com/packit/packit-service/issues/1123