Add hardcoded reposdir to the kernel checking
To not fail under the EUS systems, we are adding the same method used in the
checking if the packages are up-to-date. When we detect that we have the
hardcoded repofiles present and it matches the running system (CL8.4, CL8.5,
OL8.4) we use them instead of the system repofiles in the repoquery command
call.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>