Use YumBase instead of YumBaseCli to get pkg updates
YumBaseCli was causing the logging of convert2rhel to be broken - after
instantiating YumBaseCli every log message was duplicate on the output.

The reason is that initialization of the YumBaseCli class instance includes
calling logging.basicConfig() which sets up handlers on the root logger.

It is possible to use the YumBase class, which does not affect
logging, to get the same data - list of packages having updates
available.

The YumBaseCli logging handlers interfered with the convert2rhel logging as
we use propage=True to be able to use caplog in our unit tests:
https://github.com/oamg/convert2rhel/pull/179