Exclude iwlax2xx-firmware dependency from the conversion in OL8 (#483)
* Exclude iwlax2xx-firmware dependency from the conversion in OL8

This package is causing problems when replacing the package `iwl7260-firmware`
to it's RHEL counterpart.  The reason for that is in OL8 they splitted the
package `iwl7260-firmware` and the `iwlax2xx-firmware` the later became a
dependency of the former. In RHEL this is not true, there is only the package
`iwl7260-firmware` which contains all the files.

Jira reference: https://issues.redhat.com/browse/RHELC-567

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>