Remove old rhn* packages from OL8 config
The rhn* packages are causing problems during the conversion in OL8.4 (EUS),
since in the EUS version we are not enabling the appstream repositories (as
OracleLinux does not make appstream public for older versions) we were facing
problems with these specific packages.

Since the rhn* packages were used in the past in conjunction with Satellite 5,
and both are EOL, we decided to remove this package as it does not affect the
conversion in any way and it should not be used.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>