Disable new checks for OL8.latest
We have decided to disable the checks `is_installed_kernel` and
`check_package_updates` for OL8.6 and below, since we can't make sure that the
system will be in a state of conformity every time we run the conversion.

The reason for that, is that every OL distribution has enable by default the
repository `ol8_baseos_latest`, which will have the latest packages available
when a new release is made public (8.6, 8.7, 8.8 and so on...)

Since we can't guarantee the latest OL release, we voted to just disable the
checks for any version that is below OL8.latest, that means, for now, that the
latest release is OL8.6.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>