Set gpgcheck to one for ubi repos.
We setup UBI and centos6 package repositories for downloading the
subscription-manager packages.  It is good practice to set gpgcheck=1
for any yum repositories.  gpgcheck=0 shouldn't be exploitable since
we're hardcoding the https:// urls to the ubi repos on our CDN but it is
part of a good defense in depth.