Try setting the commit that codecov sees to the github_sha.
For pull_request events, github is merging the PR to the current HEAD of
the main branch.  It looks like codecov doesn't account for this.  So it
is looking at the parent of the PR's code.  Doing that will causes
problems:

* The coverage of the BASE will be different on the two branches.
* The line numbers won't match up in the PR's HEAD and the merged branch
  HEAD.

Try to tell codecov to use the merged branch in order to workaround
this.