Fix setup playbook feature allowing user to skip os version check.
As reported in #532, the logic in the ansible playbook was to fail when
either the version check showed an unhandled version OR the
skip_os_version_check was set to true. This meant that the toggle would
either check the version or the playbook will fail.

By using AND in that logic, we get the behaviour we want.  Either the
toggle will check the version or the playbook's check will succeed

It is confusing because it is a `fail:` task.  So a true value in the
conditional WILL fail the playbook while a false value in the
conditional means that the playbook WILL NOT fail here.

Fixes #532