Refactor the way we handle the yum/dnf transaction.
This commit introduces a rework of the way we handle the transactions
for the yum and dnf package managers.

Mainly, we are introducing 3 new classes for it. An abstract class that
handles the common/default interfaces for the transaction classes, a Yum
transaction handler class and another Dnf transaction handler class to
deal with the specific code that each of them needs to handle.

The main reason to do this split, is that we want to check for any
transaction failures before we actually process any packages after the
point of no return. To do that, we need to persist the transaction/class
object across the session (using classes).

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>