Workaround for yum i18n startup
We are adding a new initialization module to wrap the main function calls and
initialize things in an early step before actually running the conversion.

This is mainly needed because yum is loaded before we can actually setup the
locale for the system, so, we are wrapping up the main to run after we set the
locale.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>