[RHELC-570] Cleanup pexpect terminal size (#475)
* Set terminal size before pexpect spawns the process

Certain programs like yumdownloader need to have the terminal size set
before they start.  The version of pexpect in RHEL8 has a way,
pexpect.spawn()'s dimensions parameter, to set the terminal size at that
time.  We should make use of it there.

On RHEL7, pexpect doesn't have an official way to set the window size
but we have been using a workaround which involves overriding
setwinsize() to use different values since the RHEL7 version of pexpect
calls setwinsize() to set its own hardcoded values.  This patch cleans
that usage up slightly so that setwinsize is overridden when __init__()
is run but reverted back to its normal value afterwards.

* Test that an unknown argument other than dimensions raises a TypeError

Our pexpect-2.3 forwards-compatibility code catches TypeError so that it
can tell whether to emulate dimensions.  Test that when the TypeError
was for a different unknwon argument, the original TypeError is
re-raised

* Fix testing of the compat code for the pexpect dimensions argument

The compat code was not being tested before.  Why that was has several,
seemingly unrelated pieces, though, so this expanation is going to take
a bit.

* By default, pytest captures stdout so that it can display it only when
  relevant (a failing test) and so that tests can assert on what was
  sent to stdout.

* This feature causes pexpect-2.3's setting of the window size when the
  program is spawned to raise an exception.  Using pytest capture,
  setting the window size only works after the subprocess has been
  spawned, not before.

* We don't see the exception because pexpect-2.3 unconditionally catches
  any exception from setwinsize() in the place it is called during
  startup.  The pexpect code says that being unable to set the window
  size at startup is a problem but not a showstopper so that's why it's
  ignoring any errors.

  In order to find exceptions from setwinsize() at startup, I had to
  instrument the pexpect code, adding something like this to _spawn()::

    try:
        self.child_fd = sys.stdout.fileno() # used by setwinsize()
        self.setwinsize(24, 80)
    except:
+       import traceback
+       log = open('/tmp/pexpect.log', 'a')
+       log.write(traceback.format_exc())
+       log.close()
        # Some platforms do not like setwinsize (Cygwin).
        # This will cause problem when running applications that
        # are very picky about window size.
        # This is a serious limitation, but not a show stopper.
        pass

* In our code, right after we create the pexpect-managed process, we
  called setwinsize() again because setting the window size was
  unreliable.  For programs which didn't check the window size after
  they start (RHEL7's yumdownloader), that probably would be too late
  but it would work if there was a better behaved program which was
  experiencing this issue.

* However, setting the window size again meant that unittests which were
  checking the size weren't seeing what the program would see at
  startup; they were seeing the setting that happened after the program
  was already running.  This meant that the unittests were not failing
  but they weren't testing the case that is problematic for
  yumdownloader either.

How am I fixing this?

* Change the unittests to disable capfd.  This makes the exception that
  is the root cause go away. (Note: if you work on the setwinsize code,
  you could introduce a different exception.  That would need to be
  diagnosed just like this was, instrumenting the pexpect code to see if
  you are getting a traceback there.)

* Continue to call setwinsize() after startup in run_command_in_pty().
  This is because we aren't quite sure where the problems with it are
  showing up.  We can revisit removing this in the future.

* Make the unittests that test that winsize is set on startup use the
  compat class, PexpectSpawnWithDimensions, directly.  That way we won't
  be confused by the later setwinsize().

Final note: The missing coverage for the single line of the setwinsize
function is still missing even with all these changes.  However, I have
tested with extra logging inside of that function so I know that the
function is called and returns without errors now.  It seems like this
is a bug in python coverage.  I've tried to make a more isolated test
case for reporting upstream but so far my attempts have failed....

* Remove additional setting of window size

A second setting of the pexpect pseudo-terminal window size shouldn't
be necessary anymore because we set the size properly now when
instantiating the pexpect.spawn class.

* Fix filehandle closed errors when using logging with capsys disabled.

The problem is that creating logger handles while pytest capture is
active will set those handlers to write to pytest's capture buffers
instead of to stdout and stderr.  If a later test disables capture, any
logging calls made during that test will attempt to write to a closed
buffer.  Fix this by removing the logger handlers created in
logger_test.py

* Create a new fixture, clear_loggers.
* Use this fixture on test_logger_handlers() and test_tools_opts_debug()
  which use capsys and call `logger_module.setup_logger_handler()` which
  creates handlers.
* Also switch those functions to use the global_tool_opts fixture.  I
  noticed that they were modifying tool_opts and not setting it back to
  its initial state which could cause unexpected effects in another
  test so using global_tool_opts is cleaner.

Co-authored-by: Michal Bocek <mbocek@redhat.com>